pyproject.toml
[project]
name = "gnss-tid"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "astropy>=6.1.6",
    "bokeh>=3.7.3",
    "cartopy>=0.24.1",
    "dask[distributed]==2025.7",
    "datashader>=0.18.2",
    "graphviz>=0.21",
    "h5netcdf>=1.4.1",
    "h5py>=3.12.1",
    "hvplot>=0.12.0",
    "hydra-core>=1.3.2",
    "hydra-joblib-launcher>=1.2.0",
    "jupyter>=1.1.1",
    "jupyter-bokeh>=4.0.5",
    "jupyterlab>=4.3.0",
    "lz4>=4.4.4",
    "matplotlib>=3.9.2",
    "metpy>=1.6.3",
    "numba>=0.61.2",
    "numcodecs>=0.16.0",
    "numpy>=2.1.3",
    "pymap3d>=3.1.0",
    "scikit-image>=0.24.0",
    "scikit-learn>=1.5.2",
    "scipy==1.15.2",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.5",
    "torch>=2.5.1",
    "tqdm>=4.67.0",
    "tqdm-joblib>=0.0.5",
    "xarray==2025.7.1",
    "zarr>=3.0.7",
]

[project.scripts]
tid = "gnss_tid.main:main"

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["src/gnss_tid"]
